+++ R/plot_kfold_cv.R
@@ -2,8 +2,8 @@
 #' @name plot_kfold_cv
 #'
 #' @description This function plots the aggregated residuals of k-fold cross-validated
-#'                models against the outcome. This allows to evaluate how the model performs
-#'                according over- or underestimation of the outcome.
+#'   models against the outcome. This allows to evaluate how the model performs
+#'   according over- or underestimation of the outcome.
 #'
 #' @param data A data frame, used to split the data into \code{k} training-test-pairs.
 #' @param formula A model formula, used to fit linear models (\code{\link[stats]{lm}})
@@ -45,14 +45,6 @@
 #' plot_kfold_cv(efc, fit = fit)
 #'
 #' @import ggplot2
-#' @importFrom datawizard data_partition
-#' @importFrom dplyr mutate ungroup summarise
-#' @importFrom purrr map map2
-#' @importFrom tidyr unnest
-#' @importFrom graphics plot
-#' @importFrom stats as.formula formula family poisson glm lm predict
-#' @importFrom purrr map
-#' @importFrom MASS glm.nb
 #' @export
 plot_kfold_cv <- function(data, formula, k = 5, fit) {
   # make sure that data is a data frame
@@ -74,7 +66,7 @@ plot_kfold_cv <- function(data, formula, k = 5, fit) {
     else
       fam <- NULL
   } else {
-    stop("Either `formula` or `fit` must be supplied.", call. = F)
+    stop("Either `formula` or `fit` must be supplied.", call. = FALSE)
   }
 
   # get name of response variable and get variable label, if
@@ -90,7 +82,7 @@ plot_kfold_cv <- function(data, formula, k = 5, fit) {
       # pair, get deviance residuals and response value
       kfolds <- do.call(rbind, lapply(1:k, function(i) {
         out <- datawizard::data_partition(data, training_proportion = .8)
-        data.frame(train = I(list(out$training)), test = I(list(out$test)))
+        data.frame(train = I(list(out[[1]])), test = I(list(out$test)))
       }))
       res <- kfolds %>%
         dplyr::mutate(model = purrr::map(.data$train, ~ stats::glm(formula, data = .x, family = stats::poisson(link = "log")))) %>%
@@ -102,7 +94,7 @@ plot_kfold_cv <- function(data, formula, k = 5, fit) {
       # pair, get deviance residuals and response value
       kfolds <- do.call(rbind, lapply(1:k, function(i) {
         out <- datawizard::data_partition(data, training_proportion = .8)
-        data.frame(train = I(list(out$training)), test = I(list(out$test)))
+        data.frame(train = I(list(out[[1]])), test = I(list(out$test)))
       }))
       res <- kfolds %>%
         dplyr::mutate(model = purrr::map(.data$train, ~ MASS::glm.nb(formula, data = .))) %>%
@@ -119,7 +111,7 @@ plot_kfold_cv <- function(data, formula, k = 5, fit) {
     # train data
     kfolds <- do.call(rbind, lapply(1:k, function(i) {
       out <- datawizard::data_partition(data, training_proportion = .8)
-      data.frame(train = I(list(out$training)), test = I(list(out$test)))
+      data.frame(train = I(list(out[[1]])), test = I(list(out$test)))
     }))
     res <- kfolds %>%
       dplyr::mutate(model = purrr::map(.data$train, ~ stats::lm(formula, data = .))) %>%