+++ NEWS.md
@@ -1,3 +1,10 @@
+# sjPlot 2.8.10
+
+## General
+
+* Minor changes to work properly with updates of downstream packages and 
+  various smaller bug fixes and improvements.
+
 # sjPlot 2.8.9
 
 ## Bug fixes
@@ -130,7 +137,7 @@
 
 ## General
 
-* Standardization of model parameters (in `plot_model()` or `tab_model()`) now uses standardization based on refitting the model (see [vignette](https://easystats.github.io/effectsize/articles/standardize_parameters.html) for details).
+* Standardization of model parameters (in `plot_model()` or `tab_model()`) now uses standardization based on refitting the model (see [vignette](https://easystats.github.io/parameters/articles/model_parameters_standardized.html) for details).
 
 ## Changes to functions